chore: add docs for more screens and move vote button logic inside
