+++ docusaurus/docs/reactnative/ui-components/poll-answers-list.mdx
@@ -10,6 +10,8 @@ A component used to render a list of answers that a `Poll` has. The results will
 
 It will render its default `PollAnswersListContent`, which can be overridden for custom UI. Its children have access to the entire `poll` state through the `usePollState` hook.
 
+To access the paginated list of answers, the `usePollAnswersPagination` hook can be used.
+
 ## Props
 
 ### <div class="label description">_forwarded from [MessageContext](../../contexts/message-context#message)_ props</div> message {#message}
