+++ docusaurus/docs/reactnative/ui-components/poll-option-full-results.mdx
@@ -2,3 +2,55 @@
 id: poll-option-full-results
 title: PollOptionFullResults
 ---
+
+import MessageProp from '../common-content/contexts/message-context/message.mdx';
+import Poll from '../common-content/contexts/poll-context/poll.mdx';
+
+A component all of the votes for a specific `Poll` option. The results will be paginated and updated as they change. Needs to be structured inside a [`Channel` component](../core-components/channel).
+
+It will render its default `PollOptionFullResultsContent`, which can be overridden for custom UI. Its children have access to the entire `poll` state through the `usePollState` hook.
+
+To access the paginated list of votes, the `usePollOptionVotesPagination` hook can be used.
+
+## Props
+
+### <div class="label description">_forwarded from [MessageContext](../../contexts/message-context#message)_ props</div> message {#message}
+
+<MessageProp />
+
+### `poll` *
+
+<Poll />
+
+### `option` *
+
+The `poll` option that we want to display the results for.
+
+| Type   |
+| ------ |
+| object |
+
+### `additionalFlatListProps`
+
+A prop used to override the underlying [`FlatList`](https://reactnative.dev/docs/flatlist#props) props of the `PollOptionFullResults`.
+
+```jsx
+const flatListProps = { bounces: true };
+
+<PollOptionFullResults additionalFlatListProps={flatListProps} />;
+```
+
+| Type   |
+| ------ |
+| object |
+
+### `PollOptionFullResultsContent`
+
+A `Component` prop used to render the content of the `PollOptionFullResults` component.
+
+The component has full access to the entire `Poll` reactive state through the `usePollState` hook.
+
+| Type          | Default                                        |
+| ------------- | ---------------------------------------------- |
+| ComponentType | [`PollOptionFullResultsContent`](https://github.com/GetStream/stream-chat-react-native/blob/main/package/src/components/Poll/PollOptionFullResults.tsx) |
+
